Move export handling to VisibilityChecker (#9342)
* no-op: Remove unused `firstSymbol` parameter

* no-op: recursiveExportSymbol -> recursiveSetIsExported

* no-op: Accept `isExported` bool in `recursiveSetIsExported`

This way, we can use this function for recursively setting or unsetting
the `isExported` flag on a symbol.

* no-op: Isolate base case and recursive case

Better ignoring whitespace.

The old version of `recursiveSetIsExported` unfolded the recursive case
one level. That was ~fine before, when logic that needs to be done at
each level is trivial (e.g., `flags.isExported = true`).

But I plan to add more logic here, and I didn't want to have to
duplicate it in both the base and recursive cases.

* no-op: Only one switch statement over symbol kind

Looks better ignoring whitespace

* no-op: gs-> ctx

* prework: Add unsetAllExportedInPackage

By moving from packager.cc to VisibilityChecker.cc, duplicate export
checking will now run on the fast path.

The new algorithm, instead of being powered by sorting a deduplicating
`FullyQualifiedName` objects for the exports, is going to be based on
whether a symbol is **already** exported by the time an `export` line
marks it as exported.

That means that we have to start from a starting point where nothing in
the package is exported, so here's a helper that does that.

* prework: Thread some variables through

* prework: Stop using FullyQualifiedName in addExport

* Move export handling to VisibilityChecker

* Remove FullyQualifiedName

* Put the autocorrect suggestion back

* Code review comments