Simplify NamedLiteralType
I noticed that there were both `asName` and `unsafeAsName`, and I wasn't
sure what the difference was.

Turned out: they're identical. And also, in various places there were
still switch statements that were "unconditional," likely a relic from
back when there was only one LiteralType (now there are three, for
performance).

I took the liberty of simplifying some code.

I opted to delete the `asName()` function too, because it was asserting
that `literalKind` was either `Symbol` or `String`, but those are the
only two values left in the `LiteralKind` enumeration.