Make "bad parameter" errors localized to the error (#9392)
* Add test showing old behavior

* fastmod 'Unknown argument name' 'Unknown parameter name'

* Show error line for method definition parameters

* fastmoda 'specified for argument' 'specified for parameter'

* Use the tree locs for errors, not the symbol locs

The ParamInfo's loc for a method definition gets updated the first time
that its type is set, which moves the loc of the ParamInfo for `name`
from a loc in the MethodDef declLoc to a loc in the signature. When
parsing the second signature, since it can't find the parameter, it was
printing the location using `param.loc`, which came from the symbol
table, and thus was pointing to the first location, where `name`
successfully had been defined.

These fillInInfoFromSig errors should be more local/syntactic, so now
all the errors we report are from the current signature or MethodDef
node, instead of from the symbol table (though we do use the location
from the symbol table to figure that out, c.f. the `.contains` call
below)

* Show changes to the test

* Fix a bug with `<blk>` params

* Some cli tests get better