Also handle BlockPass correctly
A note about the expandToLeadingComma helper:

If there's a leading comma, it can't be `<something>, { ... }` because
that never parses. Which means if there's a leading comma, it's from a
block pass `<something>, &blk`.

Also, the `expandToLeadingComma` works even to also just remove a
leading space (it's kind of a misnomer) which means that it will
actually remove the space in `foo() { ... }`.

Previously we got this "for free" by how the `blockLoc()` helper worked.