New kernel_raise.rb are all wrong
These need to be called via `raise MyError, {input: ""}`.

At least we're not generating something syntactically invalid.
Trying to figure out that we're in this weird context from within the
"missing required kwargs" error is not going to be easy.

So I'm going to call this close enough, and let it be for now.

I'm especially not concerned because I think that custom exception
classes basically have to take a positional `"..."` parameter, so in
some sense these that accept required kwargs are already problematic (to
the Ruby VM).