Prevent autocorrecting foo(&:bar)
Sorbet expands `foo(&:bar)` to

    foo do |*tmp|
      tmp[0].bar(*tmp[1, INT_MAX])
    end

If `bar` has required keyword arguments, then you'd need to do something
like

    foo do |*tmp|
      tmp[0].bar(*tmp[1, INT_MAX], req1:, req2:)
    end

In other parts of Sorbet we have an autocorrect that does actually
expand the `foo(&:bar)` to this longer syntax, but for this case of
missing keyword args, I would prefer not to for the time being. Instead,
let's just abort the autocorrect.