Track blockLoc explicitly for better errors/fixes (#9411)
* Show bad behavior before

* Track blockLoc explicitly

It's annoying that we have to recompute it by looking at the source. We
should just pay the extra bytes to allocate it.

This memory only lives as long as we're typechecking a single file, so
it shouldn't meaningfully contribute to increased RSS.

* Also handle BlockPass correctly

A note about the expandToLeadingComma helper:

If there's a leading comma, it can't be `<something>, { ... }` because
that never parses. Which means if there's a leading comma, it's from a
block pass `<something>, &blk`.

Also, the `expandToLeadingComma` works even to also just remove a
leading space (it's kind of a misnomer) which means that it will
actually remove the space in `foo() { ... }`.

Previously we got this "for free" by how the `blockLoc()` helper worked.

* Show tests after

* Other tests that change