First-class KeywordArgResponse handling
Handled by calls.cc, so that we don't have reimplement that ad hoc in
hover.cc. Will make it easier if we want to make use of this in other
requests.

It should be pretty easy to use this to implement support for go-to-def
and go-to-type-def. If we wanted to add support for references, we would
need to also build support for a `PARAM_INFO` kind of `lsp::Query`.
Right now, references only lets you find references of symbols, unclear
whether that's something worth doing right now.