One test changes
I thought about this for a while. What's happening is that the loc of
the keyword symbol and the synthetic variable of the value are
different:

    example(account:)
    #       ^^^^^^^ the `:account` symbol
    #       ^^^^^^^^ the `account` variable

~~These locs are load bearing.~~ Not sure whether these locs are load
bearing at the moment, I had done some spelunking but now I think that
it might not matter if we wanted to change them to make them equal.

Since the `:account` symbol is a more narrow loc, it gets sorted to
appear lower in the `queryResponses`, and thus we hit the `isKeywordArg`
case instead of hitting the `isIdent` case.

I thought about ways to recover this (change the loc of the symbol as
mentioned above, or do some more "scan back over the `queryResponses`"
logic like we used to be doing before this PR).

But then I thought about it, and I'm not even sure what the user would
want to see here. It's possible that the answer is "both," it's possible
that the answer is just the keyword param info. And since it's
ambiguous, I think that asking people to manually un-pun the keyword arg
to get the one that they're after is not that bad.

Anyways, out of convenience, I've opted to just let the behavior change
here. We could change it back if that ends up being people's preference.