Change loc of punned keyword args
It looks like the original "drop the trailing :" came from Dmitry's
change to overhaul how we store arguments in the symbol table:

b7d055453361df1597fdf348c1a7434522bf01d3

The original PR is Stripe-internal: http://go/si/sorbet/pull/1888

During that PR, locations of keyword parameters became load bearing in
certain cases, because Sorbet would read the original source of the file
to show parameter names in error messages sometimes.

It looks like that helper was then cargo culted to implement punned
keyword arguments. Sorbet's `pair_keyword` builder method was used in
multiple places, which included parameter lists, argument lists, hash
lists, etc.

But this `pair_label` builder method is only used outside of method
parameter lists, so it cannot affect that use case, and we can go back
to the original location information here.

It's possible that we want to refactor things so that `pair_keyword`
uses this same loc that `pair_label` uses when it's not being used in a
parameter list, but it doesn't look like we have any tests that require
us to do that at the moment, so I'm going to punt on it.