Add autocorrect to insert missing kwargs (#9419)
* Add test showing before

* Refactor to a single error

* Add autocorrect to insert missing kwargs

* stog //test/cli:update_arity-messages

* New kernel_raise.rb are all wrong

These need to be called via `raise MyError, {input: ""}`.

At least we're not generating something syntactically invalid.
Trying to figure out that we're in this weird context from within the
"missing required kwargs" error is not going to be easy.

So I'm going to call this close enough, and let it be for now.

I'm especially not concerned because I think that custom exception
classes basically have to take a positional `"..."` parameter, so in
some sense these that accept required kwargs are already problematic (to
the Ruby VM).

* Existing tests that change

* Prevent autocorrecting foo(&:bar)

Sorbet expands `foo(&:bar)` to

    foo do |*tmp|
      tmp[0].bar(*tmp[1, INT_MAX])
    end

If `bar` has required keyword arguments, then you'd need to do something
like

    foo do |*tmp|
      tmp[0].bar(*tmp[1, INT_MAX], req1:, req2:)
    end

In other parts of Sorbet we have an autocorrect that does actually
expand the `foo(&:bar)` to this longer syntax, but for this case of
missing keyword args, I would prefer not to for the time being. Instead,
let's just abort the autocorrect.

* Actually, wrap with parens

* Add test showing the new addErrorLine calls

* Show count of missing required keyword args