calls.cc: Simplify calculating keyword arg locs (#9426)
It looks like the big scary comment about `argInBounds` might have been
written before we were tracking this `kwargLocs` mapping.

I believe that the way that mapping is handled subsumes the
`argInBounds` logic, because if the thing that the comment was talking
about is true, then there won't be an entry in the `kwargLocs` mapping,
because it's only populated when there were actual key+value pairs, not
a shape or a splat.