completion: Put the `(|)` and `{|}` back into the snippet
Follow up to #9425. After typing out a bunch of snippets, I'm noticing
how much muscle memory I had for the `()` getting put into the snippet
for me.

I noticed that IntelliJ also had this behavior, so I think it's not too
wild for Sorbet to also do it. TypeScript doesn't do it, but JavaScript
allows passing around first-class objects, while if you want to do that
in Ruby you're going to have to do something else entirely
(`x.method(:foo)` vs `x.foo`).

I've chosen to only insert the `()` for methods that have required
parameters, because the overwhelming style in the Ruby community is to
omit trailing parens for nullary method calls (despite my petitions to
change that).

While I was here, I figured that if the method had no required
arguments, but did have a required block, then we should also go ahead
and put the user's cursor inside a `{|}`. I didn't want to combine those
(put both the `(|) {|}` in one snippet) because then we end up with the
same problem we had before, which is that the presence of a snippet that
has multiple tabstops prevents completion requests from firing.