First-class KeywordArgResponse handling (#9427)
* no-op: Use exhaustive visitor in some helpers

I think that eventually I want to replace all the usages in the various
main/lsp/requests/ implementations, but that's going to have to wait for
some other time, I'm in a rush.

* no-op: Remove getTypeAndOrigins helper

It was not used widely enough to justify itself.

* First-class KeywordArgResponse handling

Handled by calls.cc, so that we don't have reimplement that ad hoc in
hover.cc. Will make it easier if we want to make use of this in other
requests.

It should be pretty easy to use this to implement support for go-to-def
and go-to-type-def. If we wanted to add support for references, we would
need to also build support for a `PARAM_INFO` kind of `lsp::Query`.
Right now, references only lets you find references of symbols, unclear
whether that's something worth doing right now.

* One test changes

I thought about this for a while. What's happening is that the loc of
the keyword symbol and the synthetic variable of the value are
different:

    example(account:)
    #       ^^^^^^^ the `:account` symbol
    #       ^^^^^^^^ the `account` variable

~~These locs are load bearing.~~ Not sure whether these locs are load
bearing at the moment, I had done some spelunking but now I think that
it might not matter if we wanted to change them to make them equal.

Since the `:account` symbol is a more narrow loc, it gets sorted to
appear lower in the `queryResponses`, and thus we hit the `isKeywordArg`
case instead of hitting the `isIdent` case.

I thought about ways to recover this (change the loc of the symbol as
mentioned above, or do some more "scan back over the `queryResponses`"
logic like we used to be doing before this PR).

But then I thought about it, and I'm not even sure what the user would
want to see here. It's possible that the answer is "both," it's possible
that the answer is just the keyword param info. And since it's
ambiguous, I think that asking people to manually un-pun the keyword arg
to get the one that they're after is not that bad.

Anyways, out of convenience, I've opted to just let the behavior change
here. We could change it back if that ends up being people's preference.

* no-op: Also make this function exhaustive

* Change loc of punned keyword args

It looks like the original "drop the trailing :" came from Dmitry's
change to overhaul how we store arguments in the symbol table:

b7d055453361df1597fdf348c1a7434522bf01d3

The original PR is Stripe-internal: http://go/si/sorbet/pull/1888

During that PR, locations of keyword parameters became load bearing in
certain cases, because Sorbet would read the original source of the file
to show parameter names in error messages sometimes.

It looks like that helper was then cargo culted to implement punned
keyword arguments. Sorbet's `pair_keyword` builder method was used in
multiple places, which included parameter lists, argument lists, hash
lists, etc.

But this `pair_label` builder method is only used outside of method
parameter lists, so it cannot affect that use case, and we can go back
to the original location information here.

It's possible that we want to refactor things so that `pair_keyword`
uses this same loc that `pair_label` uses when it's not being used in a
parameter list, but it doesn't look like we have any tests that require
us to do that at the moment, so I'm going to punt on it.

* Revert "One test changes"

This reverts commit 371494eb6fd427e3297446b638af7d21d1b34e9b.

* Update CheckSize