lsp: Report location-less error info in Diagnostic message (#9299)
* lsp: Push ErrorSection's header into first RelatedInformation

This translates to less noise in the language client.

There is an impedance mismatch in the structure that Sorbet uses to
report errors and the structure that the LSP spec needs. Because the two
structures are not equivalent, there will always be some wonkiness.

@damolina-stripe reported a particularly bad case:

```json
{
  relatedInformation: [
    {location: ..., message: ""}
    {location: ..., message: "Defined here"}
    {location: ..., message: "Note:"}
    {location: ..., message: "Please run gen-packages to fix this"}
  ]
}
```

We could attempt to solve the problem for related information entries
that are exactly "defined here" or "note" but it would be kind of
heuristic-y and weird.

In an ideal world, we might update the data structure that we use for
reporting errors to the LSP client to make it track information that
would instruct how ErrorReporter is meant to translate Sorbet's errors
into LSP diagnostics.

Another thing we might want to do in the future is to append any
ErrorLines without location information (e.g., `Note:` lines) into the
Diagnostic's top-level message, instead of into the RelatedInformation
struct (so that we don't have to pick an arbitrary location).

For the time being, at least we can cut down on the number of things in
the RelatedInformation array by flushing section headers in front of the
first ErrorLine's information.

* Update rec test

* lsp: Report location-less error info in Diagnostic message

* wip

* stog //test/lsp:update_incremental-lsp-changes