fixup: Add `#include "core/SymbolRef.h"`
It looks like this only compiled because this file was included in two
places, and those places already had an `#include` which defined
`SymbolRef` earlier in the list of `#include`'s.

I noticed this when I opened the file in my IDE and got a bunch of
errors for it.