Make Symbol queries search for multiple symbols (#9437)
* fixup: Add `#include "core/SymbolRef.h"`

It looks like this only compiled because this file was included in two
places, and those places already had an `#include` which defined
`SymbolRef` earlier in the list of `#include`'s.

I noticed this when I opened the file in my IDE and got a bunch of
errors for it.

* prework: Tidy up the Query interface

Don't use references when the types are trivially copyable `Ref`
classes.

Use `MethodRef` in place of `SymbolRef` wherever relevant.

* prework: Add a CheckSize assertion

* prework: Use std::variant for `lsp::Query`

We were kind of already doing this by way of a `Kind` tag and the
overloaded meanings of each field.

Note that this doesn't increase the size of the `lsp::Query` class,
though this is not particularl memory sensitive.

* no-op: Convert private methods to helper funcitons

I wanted to modify this, and was surprised to see that I had to edit the
header. This function doesn't use any of the task state.

* Make Symbol queries search for multiple symbols

There are a bunch of places where we're doing multiple typechecker
operations to search for mutiple SymbolRefs.

Instead of doing that, let's do one typechecker operation, but with a
query that allows matching against multiple symbols.