Only query for multiple symbols at a time (#9438)
* prework: Solve MoveMethod edge case differently

I want to remove `originalSymbol`. It was added for this check. There
are tests that fail if I comment this check out:

    //test:test_LSPTests/testdata/lsp/code_actions/move_method/class_func
    //test:test_LSPTests/testdata/lsp/code_actions/move_method/either

Looking at the failures, I think we can solve this by doing a simpler
check that doesn't require knowing the `originalSymbol`.

* no-op: Remove originalSymbol

Now that we're not using it in MoveMethod, we can get rid of this.

Also, it's a better interface: not all rename requests come from
renaming symbols, so it was not actually possible to rely on the symbol
existing (e.g., in the LocalRenamer implementation).

* Only one typechecker operation when doing renames

* no-op: clang-format

* Draw the rest of the own

Makes it so that you can only query for multiple symbols

There are actually not that many places that are only searching for one
symbol.

* Also make only one query per send

Previously, we would make one find-all-references request per
`DispatchComponent`. Now, we only make one.

* no-op: Get rid of now-unused priorRefs arg

* no-op: Remove unused variable

* no-op: Some Renamer refactors

- No need to use `shared_ptr`: the ownership is clear
- Removes `addSymbol` from public interface. Not all renamers deal with
  symbols (e.g., the LocalRenamer), so let's move this into the
  constructor for each renamer.
- As a result, don't need to take symbol in `getEdits` function, because
  we can assume that the symbol queue is populated.

* Make MethodRenamer operate on multiple symbols

* Renaming one symbol might enqueue another symbol