Drop `root` parameter from `isSubclassOrMixin` (#9451)
I want to make a version of `getSubclassesSlow` that takes multiple
roots and searches for subclasses of all of them. Rather than needing to
thread something like `UnorderedSet<ClassOrModuleRef>` to
`isSubclassOrMixin`, I think we can just use the existing `memoized`
vector, because the first thing that we do in `getSubclassesSlow` is
populate `memoized[root.id()] = true`.