Silence "Splats are only supported" error for `f(...)` (#9450)
* Add test showing current behavior

* Silence "Splats are only supported" error for `f(...)`

This is a bit of a hack. Right now, using `...` is particularly
punishing: it's frequently used by low-level code that doesn't actually
benefit from tracking types on the method call, and then the user gets
an error saying "Splats are only supported where the size of the array
is known statically" to which the only answer is downgrade the file or
tack on a T.unsafe, assuming you can't refactor the code to pass
arguments explicitly. In performance-sensitive code, even using
`T.unsafe` sometimes might slow down a piece of code, and it's just
noise.

Eventually, we might want to have first-class support for argument
forwarding, but in the mean time, let's just ignore it.

I've chosen to not have this error apply to the `f(*)` case for the time
being, out of an abundance of caution.

* Update a few other tests