Leave the original in the tree
This fixes the references. Another way to solve this might be to make
the DefLocSaver not care whether the symbol has an original or not. But
it seems weird that the bounds vs no bounds case produces a different
structure in the `Assign` tree, so I've opted to stop dropping the
`original` from the constant lit.