Fix find all references for type members with bounds (#9453)
* Show behavior before

* Leave the original in the tree

This fixes the references. Another way to solve this might be to make
the DefLocSaver not care whether the symbol has an original or not. But
it seems weird that the bounds vs no bounds case produces a different
structure in the `Assign` tree, so I've opted to stop dropping the
`original` from the constant lit.

* Report invalid type args in `# typed: false` even if there are bounds

We were short circuiting out of the function in this case, but we
shouldn't have been--by letting the fucntion run to completion, we
report a `# typed: false` invalid syntax error.

* Improve bad bounds error