Move this comment to where it seems more relevant
The old comment location is talking about the method's owner but is not
doing any logic that deals with the method owner. My guess is that at
some point it was refactored such that the `method_to_key` method got
separated from the comment documenting it, but if that's the case, it
must have been before sorbet-runtime was vendored into the Sorbet repo,
because the separation is there in the first commit adding
sorbet-runtime.