no-op: Reorganize some code in `_methods.rb` (#9456)
* Move this comment to where it seems more relevant

The old comment location is talking about the method's owner but is not
doing any logic that deals with the method owner. My guess is that at
some point it was refactored such that the `method_to_key` method got
separated from the comment documenting it, but if that's the case, it
must have been before sorbet-runtime was vendored into the Sorbet repo,
because the separation is there in the first commit adding
sorbet-runtime.

* Move all "method_to_key" methods to one spot

There were three groupings of methods that deal with signature
reflection. Now, all of these methods are in one spot in the file, which
makes it easier to have them all on screen at once.