Don't rename dispatch-related methods
Given this snippet:

```ruby
class A
  def foo; end
  #   ^ rename: bar
end

class B
  def foo; end
end

sig { params(x: T.any(A, B)).void }
def example(x)
  x.foo
end
```

Right now, Sorbet not only renames `A#foo`, but it also renames `B#foo`
just because there happened to be a successful call to `foo` on
`T.any(A, B)`.

When I polled Stripe engineers about this, I universally heard that the
behavior they preferred was that Sorbet should only rename calls to
`A#foo`, and let Sorbet report an error that now `foo_renamed` method
does not exist on `B` component of `T.any(A, B)`.

This is also the behavior that TypeScript has for renames.

The only other option in consideration was to **skip** all callsites on
unions, e.g., report that `foo` doesn't exist on `A` component of
`T.any(A, B)` after the rename. The argument against that one ended with
noticing that if you **do** want to rename all the call sites (both `A`
and `B`), you want to issue one rename on `A#foo`, one on `B#foo`, and
expect that all the call sites have been migrated. If we exclude the
union call sites, then you won't be able to programmatically rename
those: you'll just be stuck with a list of errors.