A test that changes
Before, we got an error because it was not possible for a user to
redeclare a type member in a module singleton class--writing

    class << self
      has_attached_class!(:out)
    end

would have been rejected because `has_attached_class!` could not appear
in classes, and we didn't create the implicit `<AttachedClass>` type
template in all singletons, only in `class` singletons.

Now that we're back to defining `<AttachedClass>` in module singletons,
this error shifts to pointing out the incompatible variances (because
it's no longer missing entirely).