Treat `T.attached_class` as fixed+invariant in `final!`
If a class has no subclasses, then there is no reason for the
`T.attached_class` lower bound to be `T.noreturn`: we know that the
lowest it will ever be is our current class. In that case, we can make
the bound `fixed` instead of just `upper`.

Also, since we're going to be fixing it, it may as well be invariant.
The only way this would matter would be if we could make a value with a
type like this:

    T.class_of(MyFinalClass)[... some subtype of MyFinalClass ...]

say

    T.class_of(MyFinalClass)[MyFinalClassChild]

In such a case, being invariant would be annoying, because it would mean
that

    T.class_of(MyFinalClass)[MyFinalClassChild]
      </:
    T.class_of(MyFinalClass)[MyFinalClass]

But there is no way to construct such a type, so making the type
invariant is not actually a restriction: it's exclusively more
permissive.

(I'm even thinking of a future case where one day `requires_ancestor { M
}` ties into the bound of `T.attached_class`, something like
`T.all(MyFinalClass, M)`. Even in such a future, this still doesn't
matter: a `final!` class that requires ancestors it doesn't already have
will never be able to satisfy those ancestors.)