stog //test/cli:update
I was a little worried about the change from `T.class_of(Array).new` to
`Array.new`, but it looks like that was never really intentional. In the
original change we made to delete the attached class from module
singletons, we were fine with that changing behavior:

4a0ff8dd2e1e3f4b5f52fcb9578e3c9cabd1870f

i.e., it looks like something that we might want to do something special
for, but we don't really need to stress over what that special behavior
is in this change, because it has always just been there as a "let's
capture the current behavior" test.