Remove `MethodRef::enclosingClass`
This is a vestige of when we had only a single `SymbolRef` class

If we for sure have a `MethodRef`, we know that the enclosing class is
its owner directly.

This does not delete the `SymbolRef::enclosingClass` function, which
will still still unwrap as many levels of nested ownership as required
to get the narrowest `ClassOrModule` owner.