Allow non-top-level `T.self_type`
Because `T.self_type` is modeled using a fake `T_SelfType()` type
member, and because that is defined as covariant, the polarity checking
in `variance.cc` ensures that we don't use `T.self_type` in an input
position unless the method is private.

This already got implemented in an earlier commit (e.g., when we deleted
`SelfType` and introduced `NewSelfType`, alongside using `LambdaParam`
to represent `T.self_type`), but I had intentionally not indented the
lines to show a better diff ignoring whitespace (because the lines
moved).

So as a reminder, I've included the (whitespace-only) re-indent changes
in this commit, to jog your memory about how that part of non-top-level
T.self_type works.