Make the resultTypeAsSeenFrom happen first
This matches everywhere else in the codebase.

I'm not sure if it should be this way or not, and I think that if we
unified `TypeVar` with `LambdaParam` this problem would not happen
(because it would mean that `resultTypeAsSeenFrom` and
`{approximate,instantiate}TypeVars` all combine into one.

But the idea is that as the type is being created on the way back up
from `OrType::approximateTypeVars`, we're going to combine both halves
using `Types::any` (because maybe approximating left and right could
make the `OrType` collapse). That's going to mean invoking isSubType,
and isSubType is not implemented for LambdaParam. So if we don't
eliminate all LambdaParam first, and then handle all TypeVar, then we
get an ENFORCE in approximateTypeVars.

The test that fails without this is `test/testdata/infer/result.rb`.