Exempt `badAliasMethodStub`
This method is owned by `Sorbet::Private::Static`, which is almost
certainly not what the bad alias method was called on.

If you look in `test/testdata/namer/alias_method.rb`, you see that we
have this bad method alias stub method so that if Sorbet fails to
resolve where the alias method targets, we don't report errors at calls
to it (rather, we treat it like an untyped method that accepts whatever
args).

But that means that realistically, we want to skip all the logic that
`dispatchCallSymbol` would otherwise do, including any calls to
`resultTypeAsSeenFrom`.