Carve out exception for `T.untyped` bounds
Putting `fixed: T.untyped` has the effect of basically forgetting that
there was a type member in a signature, so things like `T.nilable(...)`
would eagerly collapse.

I go back and forth on what the best thing to do would be. Another
option would be: don't allow redeclaring a `fixed: T.untyped` type
member as anything other than `fixed: T.untyped`.

But this would fly in the face of how people use `fixed: T.untyped` today,
which is to exclude most places from needing to supplying a bound, but
allowing individual classes to opt in.

Short of that, the easiest thing is just to opt `T.untyped` bounds from
this new behavior.