Implement `T::Module` (#8076)
* Start defining AttachedClass for modules again

Mostly a revert of #6883 (5c8575cc5a7429e0b0c3ca7a12b319b421ae823f),
which was originally developed in #6763.

When we deleted these symbols 2.5 years ago, we noted that the memory
savings were effectively a drop in the bucket. That seems to still be
the case today.

* Update symbol-table exp files

* stog //test/cli:update

I was a little worried about the change from `T.class_of(Array).new` to
`Array.new`, but it looks like that was never really intentional. In the
original change we made to delete the attached class from module
singletons, we were fine with that changing behavior:

4a0ff8dd2e1e3f4b5f52fcb9578e3c9cabd1870f

i.e., it looks like something that we might want to do something special
for, but we don't really need to stress over what that special behavior
is in this change, because it has always just been there as a "let's
capture the current behavior" test.

* prework: Some tests of existing behavior

* Implement `T::Module`

* no-op: Treat the has_attached_class error separately

* Only report has_attached_class for Module in strict

* Show the tests that change

* A test that changes

Before, we got an error because it was not possible for a user to
redeclare a type member in a module singleton class--writing

    class << self
      has_attached_class!(:out)
    end

would have been rejected because `has_attached_class!` could not appear
in classes, and we didn't create the implicit `<AttachedClass>` type
template in all singletons, only in `class` singletons.

Now that we're back to defining `<AttachedClass>` in module singletons,
this error shifts to pointing out the incompatible variances (because
it's no longer missing entirely).

* Update docs

* Update website/docs/class-of.md

Co-authored-by: Trevor Elliott <awesomelyawesome@gmail.com>

* Update website/docs/class-of.md

Co-authored-by: Trevor Elliott <awesomelyawesome@gmail.com>

* Update docs

---------

Co-authored-by: Trevor Elliott <awesomelyawesome@gmail.com>