Fix bug
This was mistakenly using `return nullptr`, when it needs to `break`,
which caused a crash because we would insert `nullptr` in the tree.

This caused a crash in a test, which suggests that we don't actually
have a test for the case where `includeContext` is given a block or is
not inside a describe for a `test_each` call.