Put RSpec logic behind experimental flag (#9571)
* Fix bug

This was mistakenly using `return nullptr`, when it needs to `break`,
which caused a crash because we would insert `nullptr` in the tree.

This caused a crash in a test, which suggests that we don't actually
have a test for the case where `includeContext` is given a block or is
not inside a describe for a `test_each` call.

* Add cli test showing behavior before

I checked that the error checksum matches with the testdata tests

* Make new EXPERIMENTAL section

* Thread flag through the various places

* Don't do RSpec logic unless the flag is there

* Put the assertion in the test files to get the same behavior back