Build exhaustiveness support for type members (#9535)
* Add a test showing behavior before

* Build exhaustiveness support for type members

The key observation is that if we just return the `droppedFromUpper`,
then that means that we're not returning something that's a subtype of
the input. To make it a subtype of the input, we have to use `T.all` to
remember that it started as a type member, and also that there is a more
narrow bound.

* This test improves

In #9505, I wrote:

> we can't easily fix this one, so I didn't go out of my way to make it
> work.

Little did I know that it was actually not that hard.