--gen-packages mode for imports (#9080)
* --gen-packages

* track package references

* combined autocorrect

* docs

* stop after resolving

* test

* switch to multiplexJob instead of Parallel::iterate (which uses multiplexJobWait)

* one more early continue and de-indent

* only run package reference tracking when the flag is enabled

* don't allow both --gen-packages and --lsp

* Speed up untracking package references

On the slow path, the package references sets will start off empty, so
looping over them should be fast. However, because calls to untrack and
track were interleaved, the untrack calls would encounter the entries
added by the track calls and do extra work that's not needed (since the
track calls would a file A, but the untrack call would be for some other
file B). By doing all the untracking upfront, we can make it faster.

* test for invalid CLI options

* fix non determinism in test

* keep global state param const and pass in nonConstPackageDB instead

* store unique_ptr instead of optional in the queue

* Revert "store unique_ptr instead of optional in the queue"

This reverts commit 02d2c57d3da85c520439b0dae9d82631df216c3a.