Fix failing test
The logic to skip showing type args for `T.class_of(...)` types predated
the change which made that valid type annotation syntax.

Now that users can provide explicit type arguments to `T.class_of`
types, we don't have to do anything special for `T.class_of` types in
the `useValidSyntax` case, because the version for use in errors and the
version for use in syntactically-valid locations is the same.