Collapse `T.all(AppliedType, SelfTypeParam)`
There is some history here, and it's not clear to me why the `T.all` and
`T.any` case have diverged so much.

- When Dmitry added generic classes, he built the
  `T.any(AppliedType, ...)` case to only do this if `t2` was not a
  generic class:

      if (isSubType(ctx, t2, t1)) {
          return t1;
      }
      return OrType::make_shared(t1, t2);

  And on the other side, made `T.all(AppliedType, ...)` do this if `t2`
  was not a generic class:

      return AndType::make_shared(t1, t2);

- Shortly after, nelhage wrote some code to collapse more applied types
  to bottom:

  3c9fd71fb04625a7164b25c559c0428ff5023c2e

  It makes sense there is no "dual" change in `lub` for this, because we
  ~never want lubbed types to "collapse" to top--we're happy for them to
  stay at `OrType::make_shared(t1, t2)`.

- When trevor was making user-defined generics more stable, he found an
  asymmetry in `lub`, and added `isSubType` in the other direction:

  ceb996b41be96dcf85b51ca2c8d3761ac6e4fec5

- On that PR, dmitry's intuition of "if we're changing `lub`, we should
  probably also have a 'dual' change in `glb`" kicked in, and he asked
  whether we need to change `glb`.

  The response failed to consider the `SelfTypeParam` case with bounds,
  and said "We know that it's not an AppliedType, and if it's not a
  ClassType, the result will be an AndType again," which is not wrong
  but also betrays that we could have collapsed the case in this PR.

At the risk of making the same determination again, we know that it's
only the `SelfTypeParam` case that matters here. The AppliedType,
ClassType, is_proxy_type, OrType, and AndType cases are all handled
elsewhere. TypeVar can never be a subtype of another type in the absence
of a TypeConstraint. SelfType is going away soon, and will be replaced
with something that behaves like `SelfTypeParam` (when that happens, we
will need to adjust this code).

So for performance, I've chosen to guard this with
`isa_type<SelfTypeParam>(t2)`, to avoid doing two `isSubType` calls only
to realize that it doesn't collapse.

I _think_ that we could take a similar line of reasoning to put a
similar guard in `lub`, but for a reason that I have not bothered to
understand, the order of cases we handle types is different: in `glb` we
handle `AppliedType` after `is_proxy_type`, but in `lub`: before. That's
probably not load bearing, but it makes things just complicated enough
that I don't want to bother touching `lub` for what would at best be a
performance optimization when I might be wrong.