prework: Use the untyped mode in isAlreadyASubType
In the next commit, I want to be able to use `isAsSpecificAs`, but it
exposed that we weren't using `mode` in all the places we need to be.
`T.untyped` is handled pretty early in `isSubTypeUnderConstraint`: it's
one of the first cases. But the TypeVar case is higher, so it is one
more place where we need to explicitly check `mode` for whether to treat
`T.untyped` as compatible with everything.