Fix conditional send with kwargs in Prism parser (#9611)
This commit ensures kwargs are handled correctly in the Prism parser for
conditional sends, e.g. `a&.foo(b: 1)`. The conditional send logic was
missing the step that pushes the kwargs hash into the argument list. We
could add that step here, but @amomchilov suggested we just collapse it
into the logic for regular sends.

As part of this I added a handful of test cases for conditional sends,
including one for kwargs.