Remove an ancient TODO (#9532)
* testing

* Even more ENFORCE

* Use resultTypeAsSeenFrom in resolver/

* Some more calls to resultTypeAsSeenFrom

* Tidy up ENFORCEs, make lub and glb symmetric

Before this, `Types::glb` had no handling for the `LambdaParam` case,
but `Types::lub` did. That made me suspect that there was a bug.

* Use resultTypeAsSeenFromSelf for better perf

* Make the resultTypeAsSeenFrom happen first

This matches everywhere else in the codebase.

I'm not sure if it should be this way or not, and I think that if we
unified `TypeVar` with `LambdaParam` this problem would not happen
(because it would mean that `resultTypeAsSeenFrom` and
`{approximate,instantiate}TypeVars` all combine into one.

But the idea is that as the type is being created on the way back up
from `OrType::approximateTypeVars`, we're going to combine both halves
using `Types::any` (because maybe approximating left and right could
make the `OrType` collapse). That's going to mean invoking isSubType,
and isSubType is not implemented for LambdaParam. So if we don't
eliminate all LambdaParam first, and then handle all TypeVar, then we
get an ENFORCE in approximateTypeVars.

The test that fails without this is `test/testdata/infer/result.rb`.

* prework: Remove a TODO

* Exempt `badAliasMethodStub`

This method is owned by `Sorbet::Private::Static`, which is almost
certainly not what the bad alias method was called on.

If you look in `test/testdata/namer/alias_method.rb`, you see that we
have this bad method alias stub method so that if Sorbet fails to
resolve where the alias method targets, we don't report errors at calls
to it (rather, we treat it like an untyped method that accepts whatever
args).

But that means that realistically, we want to skip all the logic that
`dispatchCallSymbol` would otherwise do, including any calls to
`resultTypeAsSeenFrom`.

* requires_ancestor breaks this