Autosuggest `true, false` for T::Boolean (#9625)
* Autosuggest `true, false` for T::Boolean

It can be an common error to write something like:
```
case foo
when Integer
 ...
when T::Boolean
```
but this won't work as T::Boolean is just a sorbet type object. The proper thing would be to do a `when true, false`

I'm not sure this is the correct place/way to do this but it seems to work. One challenge is that the desugar process converts the when clause to a `===` so at this point of the code we think we're in a `===` instead of a case statement. That said sorbet doesn't seem to error if you attempt to `=== T::Boolean` in a normal case only when its handled as part of a case statement.

* PR Feedback

add a isBoolean helper to check the TypePtr
covert the tests to `*.autocorrects.exp` tests instead of CLI tests.