Directly ban using `T.self_type` in type member bounds
This was already unusable in practice, because you'd get "is not fully
defined" type errors.

On my work-in-progress `T.self_type` branch, using `T.self_type` in type
member bounds causes invariants to be violated inside the
`processBinding` instruction for the `Cast` instruction that sets up the
type of `<self>` (because it tries to create a type like `A[LambdaParam
{ T.self_type }]`, which is not fully defined).

We do want to allow this code eventually, but it's not clear to me what
the best way to do it is, and I don't want to block building better
support for `T.self_type` on also solving this problem.

I had been hoping that the `T.self_type` branch would somehow magically
solve this, but it's clear that it makes it worse, which is a little
unlucky. Maybe it's not actually worse, maybe it's directionally
accurate and we just need to keep following that direction such that we
maintain the invariants we need to.

But in any case, this unblocks the `T.self_type` work until we can think
through what a better behavior would be here.