Fix verify_prism_regression_tests.sh script
This commit updates the `verify_prism_regression_tests.sh` script to use
the correct file names now that we renamed them. It also checks the
desugar-tree-raw output, not just the parse-tree output. I also added
missing exp files that were picked up by the script.

As part of this, I needed to update the keyword_LINE.rb test to disable
stress-incremental mode, to ensure the literal integer value for the
__LINE__ constant is correct.