Special case `Object#class` intrinsic
The change in the test is expected.

The `self.class[Elem]` does not return `T.self_type` anymore. I suppose
that we could change the `[]` intrinsic to see that if you're starting
from something with type `T.class_of(Box)[T.self_type (of Box[Elem])]`
and you explicitly apply it to `self.class[Elem]`, that we just drop the
`Elem` and leave it as `T.self_type`.

But the whole reason why I added this test case was because in the past,
`self.class.new` would get inferred to have type `Elem` being
`T.untyped` when dispatching to `initialize`, which meant that you could
screw up the argument types if you did not explicitly apply it.

In practice, I think that no one knew that you could do this to save
yourself from untyped, and in any case I don't think you _should_ have
to do this to save yourself, so I'm fine regressing the behavior of
`self.class[Elem]` so long as it means that `self.class.new` still
binds `Elem` to a `SelfTypeParam` (not `T.untyped`) when invoking the
constructor.

I spent a lot of trying trying to get the behavior here via making
`T.all` collapse types differently. There is some exploration here:
<https://github.com/sorbet/sorbet/pull/9601>

But the changes in that PR were not enough to get this test's behavior
to work out, and as mentioned in the comment below, none of my tweaks
could get it quite right either: what always ended up happening was that
something got inferred to `T.untyped` when it shouldn't have.

So I'd like to go ahead with this approach, because I have tried so many
other ways that all did not work.