Assorted no-ops in packager.cc (#9642)
* no-op: Get rid of a helper function called once

The name `findPackages` stopped making sense a while ago. Rather than
pick a better name, I just merged it into the one place calling it.

* no-op: clang-format

* Move Timers to outer functions

* no-op: Move ComputePackageSCCs to "findPackages"

This technically changes the length of the `findPackages` span, but we
already have a span for the SCC condensation so I don't think nesting it
somewhere earlier matters.

* no-op: Move "findPackages" into `buildPackageDB`

* no-op: Get rid of templated `packageRunCore`

* Remove redundant timer

There is already a `packager.validatePackagesAndFiles` immediately
below.

* no-op: clang-format

* no-op: packageRunCore → validatePackagesAndFiles

* Split the validatePackagesAndFiles into each caller

* validatePackagedFiles can use `const GlobalState &`

* no-op: Do the self-import checking in PackageSpecBodyWalk